lifinity: split into v1 & v2
